+++ Libraries/LibWeb/CMakeLists.txt
@@ -71,6 +71,7 @@ set(SOURCES
     CSS/CSS.cpp
     CSS/CSSAnimation.cpp
     CSS/CSSConditionRule.cpp
+    CSS/CSSDescriptors.cpp
     CSS/CSSFontFaceDescriptors.cpp
     CSS/CSSFontFaceRule.cpp
     CSS/CSSGroupingRule.cpp
